+++ webargs/falconparser.py
@@ -140,9 +140,9 @@ class FalconParser(core.Parser):
             "Parsing files not yet supported by {0}".format(self.__class__.__name__)
         )
 
-    def handle_error(self, error, req, schema):
+    def handle_error(self, error, req, schema, error_status_code, error_headers):
         """Handles errors during parsing."""
-        status = status_map.get(error.status_code)
+        status = error_status_code or status_map.get(error.status_code)
         if status is None:
             raise LookupError("Status code {0} not supported".format(error.status_code))
         raise HTTPError(status, errors=error.messages)